+++ R/utilities-figures.R
@@ -236,9 +236,9 @@ createEsqlabsExportConfiguration <- function(projectConfiguration) {
 #' @return A list of `ggplot` objects
 #'
 #' @export
-plotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguration) {
-  validateIsOfType(observedData, DataSet)
-  validateIsOfType(projectConfiguration, ProjectConfiguration)
+createPlotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguration) {
+  validateIsOfType(observedData, "DataSet", nullAllowed = TRUE)
+  validateIsOfType(projectConfiguration, "ProjectConfiguration")
   dfDataCombined <- readExcel(projectConfiguration$plotsFile, sheet = "DataCombined")
 
   # warnings for invalid data in plot definitions from excel
@@ -246,30 +246,36 @@ plotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguratio
   missingScenarios <- setdiff(setdiff(dfDataCombined$scenario, names(simulatedScenarios)), NA)
   if (length(missingScenarios) != 0) {
     warning(messages$warningInvalidScenarioName(missingScenarios, projectConfiguration$plotsFile))
-    dfDataCombined <- dfDataCombined[!(dfDataCombined$scenario %in% missingScenarios),]
+    dfDataCombined <- dfDataCombined[!(dfDataCombined$scenario %in% missingScenarios), ]
   }
   # data set name not present in observedData
   missingDataSets <- setdiff(setdiff(dfDataCombined$dataSet, names(observedData)), NA)
   if (length(missingDataSets) != 0) {
     warning(messages$warningInvalidDataSetName(missingDataSets, projectConfiguration$plotsFile))
-    dfDataCombined <- dfDataCombined[!(dfDataCombined$dataSet %in% missingDataSets),]
+    dfDataCombined <- dfDataCombined[!(dfDataCombined$dataSet %in% missingDataSets), ]
+  }
+
+  # mandatory column label - remove rows without entry and throw error
+  missingLabel <- sum(is.na(dfDataCombined$label))
+  if (missingLabel > 0) {
+    error(messages$missingLabel())
   }
 
   # create named list of DataCombined objects
   dataCombinedNames <- unique(dfDataCombined$DataCombinedName)
-  dataCombinedList <- list()
-  for (name in dataCombinedNames) {
-    dataCombinedList[[name]] <- DataCombined$new()
+  dataCombinedList <- lapply(dataCombinedNames, \(name) {
+    dataCombined <- DataCombined$new()
     # add data to DataCombined object
     # add simulated data
     simulated <- dplyr::filter(dfDataCombined, DataCombinedName == name, dataType == "simulated")
     if (nrow(simulated) > 0) {
       for (j in 1:nrow(simulated)) {
-        dataCombinedList[[name]]$addSimulationResults(
-          simulationResults = simulatedScenarios[[simulated[j,]$scenario]]$results,
-          quantitiesOrPaths = simulated[j,]$path,
-          groups = simulated[j,]$group,
-          names = simulated[j,]$label)
+        dataCombined$addSimulationResults(
+          simulationResults = simulatedScenarios[[simulated[j, ]$scenario]]$results,
+          quantitiesOrPaths = simulated[j, ]$path,
+          groups = simulated[j, ]$group,
+          names = simulated[j, ]$label
+        )
       }
     }
 
@@ -277,36 +283,46 @@ plotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguratio
     observed <- dplyr::filter(dfDataCombined, DataCombinedName == name, dataType == "observed")
     if (nrow(observed) > 0) {
       dataSets <- observedData[observed$dataSet]
-      dataCombinedList[[name]]$addDataSets(dataSets, groups = observed$group)
+      dataCombined$addDataSets(dataSets, names = observed$label, groups = observed$group)
     }
-  }
+    dataCombined
+  })
+  names(dataCombinedList) <- dataCombinedNames
 
   # read sheet "plotConfiguration"
   dfPlotConfigurations <- readExcel(projectConfiguration$plotsFile, sheet = "plotConfiguration")
-  # remove dataCombineds that have not been generated due to missing data
+  # remove DataCombined objects that have not been generated due to missing data
   # or are not defined in sheet "DataCombined"
-  dfPlotConfigurations <- dfPlotConfigurations[dfPlotConfigurations$DataCombinedName %in% dfDataCombined$DataCombinedName,]
+  dfPlotConfigurations <- dfPlotConfigurations[dfPlotConfigurations$DataCombinedName %in% dfDataCombined$DataCombinedName, ]
   # merge limit columns to character columns xAxisLimits and yAxisLimits
-  dfPlotConfigurations <- dplyr::mutate(dfPlotConfigurations, xAxisLimits = ifelse(!is.na(xLimLower) & !is.na(xLimUpper),
-                                                                                   paste0("c(", xLimLower, ",", xLimUpper, ")"), NA),
-                                        xLimLower = NULL, xLimUpper = NULL)
-  dfPlotConfigurations <- dplyr::mutate(dfPlotConfigurations, yAxisLimits = ifelse(!is.na(yLimLower) & !is.na(yLimUpper),
-                                                                                   paste0("c(", yLimLower, ",", yLimUpper, ")"), NA),
-                                        yLimLower = NULL, yLimUpper = NULL)
+  dfPlotConfigurations <- dplyr::mutate(dfPlotConfigurations,
+    xAxisLimits = ifelse(!is.na(xLimLower) & !is.na(xLimUpper),
+      paste0("c(", xLimLower, ",", xLimUpper, ")"), NA
+    ),
+    xLimLower = NULL, xLimUpper = NULL
+  )
+  dfPlotConfigurations <- dplyr::mutate(dfPlotConfigurations,
+    yAxisLimits = ifelse(!is.na(yLimLower) & !is.na(yLimUpper),
+      paste0("c(", yLimLower, ",", yLimUpper, ")"), NA
+    ),
+    yLimLower = NULL, yLimUpper = NULL
+  )
 
   # create a list of plotConfiguration objects as defined in sheet "plotConfiguration"
-  plotConfigurationList <- apply(dfPlotConfigurations[, !(names(dfPlotConfigurations) %in% c("plotID", "DataCombinedName", "plotType"))],
-                                 1, .plotConfigurationFromRow)
+  plotConfigurationList <- apply(
+    dfPlotConfigurations[, !(names(dfPlotConfigurations) %in% c("plotID", "DataCombinedName", "plotType"))],
+    1, .plotConfigurationFromRow
+  )
 
   # create a list of plots from dataCombinedList and plotConfigurationList
   plotList <- lapply(seq_along(plotConfigurationList), \(i) {
-    dataCombined = dataCombinedList[[dfPlotConfigurations$DataCombinedName[i]]]
-    switch (dfPlotConfigurations$plotType[i],
-            individual = plotIndividualTimeProfile(dataCombined, plotConfigurationList[[i]]),
-            population = plotPopulationTimeProfile(dataCombined, plotConfigurationList[[i]]),
-            observedVsSimulated = plotObservedVsSimulated(dataCombined, plotConfigurationList[[i]]),
-            residualsVsSimulated = plotResidualsVsSimulated(dataCombined, plotConfigurationList[[i]]),
-            residualsVsTime = plotResidualsVsTime(dataCombined, plotConfigurationList[[i]])
+    dataCombined <- dataCombinedList[[dfPlotConfigurations$DataCombinedName[i]]]
+    switch(dfPlotConfigurations$plotType[i],
+      individual = plotIndividualTimeProfile(dataCombined, plotConfigurationList[[i]]),
+      population = plotPopulationTimeProfile(dataCombined, plotConfigurationList[[i]]),
+      observedVsSimulated = plotObservedVsSimulated(dataCombined, plotConfigurationList[[i]]),
+      residualsVsSimulated = plotResidualsVsSimulated(dataCombined, plotConfigurationList[[i]]),
+      residualsVsTime = plotResidualsVsTime(dataCombined, plotConfigurationList[[i]])
     )
   })
   names(plotList) <- dfPlotConfigurations$plotID
@@ -332,6 +348,7 @@ plotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguratio
   return(multiPanelPlots)
 }
 
+#' @keywords internal
 .plotConfigurationFromRow <- function(...) {
   plotConfiguration <- createEsqlabsPlotConfiguration()
   lapply(seq_along(...), function(i) {
@@ -341,11 +358,11 @@ plotsFromExcel <- function(simulatedScenarios, observedData, projectConfiguratio
       # numeric columns
       if (colName %in% c("xAxisLimits", "yAxisLimits")) {
         plotConfiguration[[colName]] <- eval(parse(text = value))
-      # character columns
+        # character columns
       } else {
         plotConfiguration[[colName]] <- value
       }
-    }}
-  )
+    }
+  })
   return(plotConfiguration)
 }